+++ ext/gmp/gmp.c
@@ -31,6 +31,7 @@
 
 /* Needed for gmp_random() */
 #include "ext/random/php_random.h"
+#include "ext/random/php_random_csprng.h"
 
 #define GMP_ROUND_ZERO      0
 #define GMP_ROUND_PLUSINF   1
