+++ ext/random/config.m4
@@ -29,4 +29,4 @@ PHP_NEW_EXTENSION(random,
       gammasection.c \
       randomizer.c,
       no,, -DZEND_ENABLE_STATIC_TSRMLS_CACHE=1)
-PHP_INSTALL_HEADERS([ext/random], [php_random.h php_random_uint128.h])
+PHP_INSTALL_HEADERS([ext/random], [php_random.h php_random_csprng.h php_random_uint128.h])
