+++ ext/random/engine_xoshiro256starstar.c
@@ -23,6 +23,7 @@
 
 #include "php.h"
 #include "php_random.h"
+#include "php_random_csprng.h"
 
 #include "Zend/zend_exceptions.h"
